refactor AW-168: rename component
make duplicate sla-section before deletion
