+++ strapi/src/middlewares/populate-sections-with-intro-and-localizations.ts
@@ -40,7 +40,7 @@ module.exports = (config, { strapi }) => {
                 },
               },
             },
-            "relations.sla-card-section": {
+            "sections.sla-card-section": {
               populate: {
                 props: true,
                 cards: {
